add coordinateInitialLoad(), _load() and _save()
moved the initial load of the page to coordinator.js
and call it from index.js

created _load() and _save() in dataStorage.js to handle
the storage of data.
Made them as general as possible to be easily able to switch to
a server-side storage solution.
Planned to separate tasks, project and parameters to reduce the
future traffic