add deleteTask() and deleteProject()
Reworked the handling of ID's with Projects.

Projects now have their own ID to prevent future double-use.
They are now referenced in the tasks by their ID instead of their name.

removeProjectFromTasks() is being called to remove a deleted project
from all tasks and reset the referenced to the General project which
is always existent at ID=1 and undeletable
(undeletable not yet implemented)

Additionally _assignUniqueID has been generalized to be used for
both tasks and projects.