add dynamic creation of view elements
initialize all views at first page load with correct dates
and save them in sessionStorage

from sessionStorage DOM-elements are created from these views.

the active view is saved in the sessionStorage and indicated by
indicateActiveView()

With the ID of the activeView we will be able to filter task-lists
by date which is initialized at each page-load aswell.